refactor: move AuthType to lib/auth/types, split extension utils
- AuthType enum relocated from lib/constants.ts → lib/auth/types.ts;
  all callsites updated to import from the canonical location
- lib/extension/utils.ts split into pure utils / svc.ts (side-effect
  fns) / hooks.ts (React wrappers); callsites updated accordingly
- Document-title side effects extracted from AppChrome and AdminChrome
  into useAppDocumentTitle / useAdminDocumentTitle in lib/app/hooks.ts
