+++ web/src/hooks/useTokenRefresh.ts
@@ -3,8 +3,7 @@
 import { useEffect, useRef } from "react";
 import { User } from "@/lib/types";
 import { NO_AUTH_USER_ID } from "@/lib/extension/constants";
-import { AuthTypeMetadata } from "@/lib/auth/types";
-import { AuthType } from "@/lib/constants";
+import { AuthType, AuthTypeMetadata } from "@/lib/auth/types";
 
 const REFRESH_INTERVAL = 600000;
 const MIN_REFRESH_GAP_MS = REFRESH_INTERVAL - 60000;
