+++ web/src/lib/extension/hooks.ts
@@ -1,0 +1,19 @@
+"use client";
+
+import { useEffect } from "react";
+import {
+  sendAuthRequiredMessage,
+  sendMessageToParent,
+} from "@/lib/extension/svc";
+
+export function useSendAuthRequiredMessage(): void {
+  useEffect(() => {
+    sendAuthRequiredMessage();
+  }, []);
+}
+
+export function useSendMessageToParent(): void {
+  useEffect(() => {
+    sendMessageToParent();
+  }, []);
+}
