+++ web/src/lib/extension/utils.ts
@@ -1,6 +1,3 @@
-import { useEffect } from "react";
-import { CHROME_MESSAGE } from "./constants";
-
 export type ExtensionContext = "new_tab" | "side_panel" | null;
 
 // Returns the origin of the Chrome extension panel (our parent frame).
@@ -27,40 +24,3 @@ export function getExtensionContext(): {
   }
   return { isExtension: false, context: null };
 }
-export function sendSetDefaultNewTabMessage(value: boolean) {
-  if (typeof window !== "undefined" && window.parent !== window) {
-    window.parent.postMessage(
-      { type: CHROME_MESSAGE.SET_DEFAULT_NEW_TAB, value },
-      getPanelOrigin()
-    );
-  }
-}
-
-export const sendAuthRequiredMessage = () => {
-  if (typeof window !== "undefined" && window.parent !== window) {
-    window.parent.postMessage(
-      { type: CHROME_MESSAGE.AUTH_REQUIRED },
-      getPanelOrigin()
-    );
-  }
-};
-
-export const useSendAuthRequiredMessage = () => {
-  useEffect(() => {
-    sendAuthRequiredMessage();
-  }, []);
-};
-
-export const sendMessageToParent = () => {
-  if (typeof window !== "undefined" && window.parent !== window) {
-    window.parent.postMessage(
-      { type: CHROME_MESSAGE.ONYX_APP_LOADED },
-      getPanelOrigin()
-    );
-  }
-};
-export const useSendMessageToParent = () => {
-  useEffect(() => {
-    sendMessageToParent();
-  }, []);
-};
