+++ web/src/views/AppPage.tsx
@@ -24,7 +24,7 @@ import { useCurrentUser } from "@/lib/users/hooks";
 import NoAgentModal from "@/sections/modals/NoAgentModal";
 import PreviewModal from "@/sections/modals/PreviewModal";
 import Modal from "@/refresh-components/Modal";
-import { useSendMessageToParent } from "@/lib/extension/utils";
+import { useSendMessageToParent } from "@/lib/extension/hooks";
 import { SUBMIT_MESSAGE_TYPES } from "@/lib/extension/constants";
 import { getSourceMetadata } from "@/lib/sources";
 import { SourceMetadata } from "@/lib/search/interfaces";
