+++ web/src/views/admin/UsersPage/UsersSummary.tsx
@@ -9,7 +9,7 @@ import Text from "@/refresh-components/texts/Text";
 import Link from "next/link";
 import { ADMIN_ROUTES } from "@/lib/admin-routes";
 import { useAuthTypeMetadata } from "@/hooks/useAuthTypeMetadata";
-import { AuthType } from "@/lib/constants";
+import { AuthType } from "@/lib/auth/types";
 import InviteOnlyCard from "./InviteOnlyCard";
 
 // ---------------------------------------------------------------------------
